+++ bindings/wisdom.h
@@ -6,7 +6,7 @@
 
 /** \mainpage Wisdom API Documentation
 
-<b>Version 0.3.3</b>
+<b>Version 0.3.6</b>
 
 Copyright (c) 2024 Ilya Doroshenko. All rights reserved.
 License: MIT
@@ -1306,9 +1306,8 @@ enum WisResourceAccessBits {
     ResourceAccessShadingRate = 1 << 15, ///< Shading rate access. Used in variable shading rate.
     ResourceAccessVideoDecodeRead = 1 << 16, ///< Video decode read access.
     ResourceAccessVideoDecodeWrite = 1 << 17, ///< Video decode write access.
-    ResourceAccessPresent = 1 << 18, ///< Present access. Used fpr swapchain presentation.
-    ResourceAccessResolveDest = 1 << 19, ///< Resolve destination access. Used in multisampling.
-    ResourceAccessResolveSource = 1 << 20, ///< Resolve source access. Used in multisampling.
+    ResourceAccessResolveDest = 1 << 18, ///< Resolve destination access. Used in multisampling.
+    ResourceAccessResolveSource = 1 << 19, ///< Resolve source access. Used in multisampling.
     ResourceAccessNoAccess = 1 << 31, ///< No access. Used to indicate no access throughout pipeline.
 };
 
@@ -2229,33 +2228,29 @@ WISDOM_API WisResult VKDeviceCreateShaderResource(VKDevice self, VKTexture textu
  * The value is used to correctly determine descriptor page alignment for descriptor buffer.
  * @param self valid handle to the Device
  * @param heap The type of the descriptor heap to get the alignment for.
- * @param alignment The alignment of the descriptor table in bytes.
- * @return Result with StatusOk on success.
- * Error in WisResult::error otherwise.
+ * @return The alignment of the descriptor table in bytes.
  * */
-WISDOM_API WisResult VKDeviceGetDescriptorTableAlignment(VKDevice self, WisDescriptorHeapType heap, uint32_t* alignment);
+WISDOM_API uint32_t VKDeviceGetDescriptorTableAlignment(VKDevice self, WisDescriptorHeapType heap);
 
 /**
  * @brief Returns the size of the descriptor buffer unit in bytes.
  * @param self valid handle to the Device
  * @param heap The type of the descriptor heap to get the unit size for.
- * @param size The size of the descriptor buffer unit in bytes. Descriptor unit is the size of one descriptor.
- * @return Result with StatusOk on success.
- * Error in WisResult::error otherwise.
+ * @return The size of the descriptor buffer unit in bytes. Descriptor unit is the size of one descriptor.
  * */
-WISDOM_API WisResult VKDeviceGetDescriptorBufferUnitSize(VKDevice self, WisDescriptorHeapType heap, uint32_t* size);
+WISDOM_API uint32_t VKDeviceGetDescriptorBufferUnitSize(VKDevice self, WisDescriptorHeapType heap);
 
 /**
  * @brief Creates a descriptor buffer object.
  * @param self valid handle to the Device
  * @param heap_type The type of the descriptor heap to create the descriptor buffer with.
  * @param memory_type The type of the descriptor memory to create the descriptor buffer with.
- * @param descriptor_count The number of descriptors to allocate in the descriptor buffer.
+ * @param size_bytes The number of bytes to allocate for the descriptor buffer.
  * @param buffer VKDescriptorBuffer on success (StatusOk).
  * @return Result with StatusOk on success.
  * Error in WisResult::error otherwise.
  * */
-WISDOM_API WisResult VKDeviceCreateDescriptorBuffer(VKDevice self, WisDescriptorHeapType heap_type, WisDescriptorMemory memory_type, uint32_t descriptor_count, VKDescriptorBuffer* buffer);
+WISDOM_API WisResult VKDeviceCreateDescriptorBuffer(VKDevice self, WisDescriptorHeapType heap_type, WisDescriptorMemory memory_type, uint64_t size_bytes, VKDescriptorBuffer* buffer);
 
 /**
  * @brief Queries if the device supports the feature.
@@ -3265,33 +3260,29 @@ WISDOM_API WisResult DX12DeviceCreateShaderResource(DX12Device self, DX12Texture
  * The value is used to correctly determine descriptor page alignment for descriptor buffer.
  * @param self valid handle to the Device
  * @param heap The type of the descriptor heap to get the alignment for.
- * @param alignment The alignment of the descriptor table in bytes.
- * @return Result with StatusOk on success.
- * Error in WisResult::error otherwise.
+ * @return The alignment of the descriptor table in bytes.
  * */
-WISDOM_API WisResult DX12DeviceGetDescriptorTableAlignment(DX12Device self, WisDescriptorHeapType heap, uint32_t* alignment);
+WISDOM_API uint32_t DX12DeviceGetDescriptorTableAlignment(DX12Device self, WisDescriptorHeapType heap);
 
 /**
  * @brief Returns the size of the descriptor buffer unit in bytes.
  * @param self valid handle to the Device
  * @param heap The type of the descriptor heap to get the unit size for.
- * @param size The size of the descriptor buffer unit in bytes. Descriptor unit is the size of one descriptor.
- * @return Result with StatusOk on success.
- * Error in WisResult::error otherwise.
+ * @return The size of the descriptor buffer unit in bytes. Descriptor unit is the size of one descriptor.
  * */
-WISDOM_API WisResult DX12DeviceGetDescriptorBufferUnitSize(DX12Device self, WisDescriptorHeapType heap, uint32_t* size);
+WISDOM_API uint32_t DX12DeviceGetDescriptorBufferUnitSize(DX12Device self, WisDescriptorHeapType heap);
 
 /**
  * @brief Creates a descriptor buffer object.
  * @param self valid handle to the Device
  * @param heap_type The type of the descriptor heap to create the descriptor buffer with.
  * @param memory_type The type of the descriptor memory to create the descriptor buffer with.
- * @param descriptor_count The number of descriptors to allocate in the descriptor buffer.
+ * @param size_bytes The number of bytes to allocate for the descriptor buffer.
  * @param buffer DX12DescriptorBuffer on success (StatusOk).
  * @return Result with StatusOk on success.
  * Error in WisResult::error otherwise.
  * */
-WISDOM_API WisResult DX12DeviceCreateDescriptorBuffer(DX12Device self, WisDescriptorHeapType heap_type, WisDescriptorMemory memory_type, uint32_t descriptor_count, DX12DescriptorBuffer* buffer);
+WISDOM_API WisResult DX12DeviceCreateDescriptorBuffer(DX12Device self, WisDescriptorHeapType heap_type, WisDescriptorMemory memory_type, uint64_t size_bytes, DX12DescriptorBuffer* buffer);
 
 /**
  * @brief Queries if the device supports the feature.
@@ -4247,26 +4238,22 @@ inline WisResult WisDeviceCreateShaderResource(WisDevice self, WisTexture textur
  * The value is used to correctly determine descriptor page alignment for descriptor buffer.
  * @param self valid handle to the Device
  * @param heap The type of the descriptor heap to get the alignment for.
- * @param alignment The alignment of the descriptor table in bytes.
- * @return Result with StatusOk on success.
- * Error in WisResult::error otherwise.
+ * @return The alignment of the descriptor table in bytes.
  * */
-inline WisResult WisDeviceGetDescriptorTableAlignment(WisDevice self, WisDescriptorHeapType heap, uint32_t* alignment)
+inline uint32_t WisDeviceGetDescriptorTableAlignment(WisDevice self, WisDescriptorHeapType heap)
 {
-    return DX12DeviceGetDescriptorTableAlignment(self, heap, alignment);
+    return DX12DeviceGetDescriptorTableAlignment(self, heap);
 }
 
 /**
  * @brief Returns the size of the descriptor buffer unit in bytes.
  * @param self valid handle to the Device
  * @param heap The type of the descriptor heap to get the unit size for.
- * @param size The size of the descriptor buffer unit in bytes. Descriptor unit is the size of one descriptor.
- * @return Result with StatusOk on success.
- * Error in WisResult::error otherwise.
+ * @return The size of the descriptor buffer unit in bytes. Descriptor unit is the size of one descriptor.
  * */
-inline WisResult WisDeviceGetDescriptorBufferUnitSize(WisDevice self, WisDescriptorHeapType heap, uint32_t* size)
+inline uint32_t WisDeviceGetDescriptorBufferUnitSize(WisDevice self, WisDescriptorHeapType heap)
 {
-    return DX12DeviceGetDescriptorBufferUnitSize(self, heap, size);
+    return DX12DeviceGetDescriptorBufferUnitSize(self, heap);
 }
 
 /**
@@ -4274,14 +4261,14 @@ inline WisResult WisDeviceGetDescriptorBufferUnitSize(WisDevice self, WisDescrip
  * @param self valid handle to the Device
  * @param heap_type The type of the descriptor heap to create the descriptor buffer with.
  * @param memory_type The type of the descriptor memory to create the descriptor buffer with.
- * @param descriptor_count The number of descriptors to allocate in the descriptor buffer.
+ * @param size_bytes The number of bytes to allocate for the descriptor buffer.
  * @param buffer WisDescriptorBuffer on success (StatusOk).
  * @return Result with StatusOk on success.
  * Error in WisResult::error otherwise.
  * */
-inline WisResult WisDeviceCreateDescriptorBuffer(WisDevice self, WisDescriptorHeapType heap_type, WisDescriptorMemory memory_type, uint32_t descriptor_count, WisDescriptorBuffer* buffer)
+inline WisResult WisDeviceCreateDescriptorBuffer(WisDevice self, WisDescriptorHeapType heap_type, WisDescriptorMemory memory_type, uint64_t size_bytes, WisDescriptorBuffer* buffer)
 {
-    return DX12DeviceCreateDescriptorBuffer(self, heap_type, memory_type, descriptor_count, buffer);
+    return DX12DeviceCreateDescriptorBuffer(self, heap_type, memory_type, size_bytes, buffer);
 }
 
 /**
@@ -5429,26 +5416,22 @@ inline WisResult WisDeviceCreateShaderResource(WisDevice self, WisTexture textur
  * The value is used to correctly determine descriptor page alignment for descriptor buffer.
  * @param self valid handle to the Device
  * @param heap The type of the descriptor heap to get the alignment for.
- * @param alignment The alignment of the descriptor table in bytes.
- * @return Result with StatusOk on success.
- * Error in WisResult::error otherwise.
+ * @return The alignment of the descriptor table in bytes.
  * */
-inline WisResult WisDeviceGetDescriptorTableAlignment(WisDevice self, WisDescriptorHeapType heap, uint32_t* alignment)
+inline uint32_t WisDeviceGetDescriptorTableAlignment(WisDevice self, WisDescriptorHeapType heap)
 {
-    return VKDeviceGetDescriptorTableAlignment(self, heap, alignment);
+    return VKDeviceGetDescriptorTableAlignment(self, heap);
 }
 
 /**
  * @brief Returns the size of the descriptor buffer unit in bytes.
  * @param self valid handle to the Device
  * @param heap The type of the descriptor heap to get the unit size for.
- * @param size The size of the descriptor buffer unit in bytes. Descriptor unit is the size of one descriptor.
- * @return Result with StatusOk on success.
- * Error in WisResult::error otherwise.
+ * @return The size of the descriptor buffer unit in bytes. Descriptor unit is the size of one descriptor.
  * */
-inline WisResult WisDeviceGetDescriptorBufferUnitSize(WisDevice self, WisDescriptorHeapType heap, uint32_t* size)
+inline uint32_t WisDeviceGetDescriptorBufferUnitSize(WisDevice self, WisDescriptorHeapType heap)
 {
-    return VKDeviceGetDescriptorBufferUnitSize(self, heap, size);
+    return VKDeviceGetDescriptorBufferUnitSize(self, heap);
 }
 
 /**
@@ -5456,14 +5439,14 @@ inline WisResult WisDeviceGetDescriptorBufferUnitSize(WisDevice self, WisDescrip
  * @param self valid handle to the Device
  * @param heap_type The type of the descriptor heap to create the descriptor buffer with.
  * @param memory_type The type of the descriptor memory to create the descriptor buffer with.
- * @param descriptor_count The number of descriptors to allocate in the descriptor buffer.
+ * @param size_bytes The number of bytes to allocate for the descriptor buffer.
  * @param buffer WisDescriptorBuffer on success (StatusOk).
  * @return Result with StatusOk on success.
  * Error in WisResult::error otherwise.
  * */
-inline WisResult WisDeviceCreateDescriptorBuffer(WisDevice self, WisDescriptorHeapType heap_type, WisDescriptorMemory memory_type, uint32_t descriptor_count, WisDescriptorBuffer* buffer)
+inline WisResult WisDeviceCreateDescriptorBuffer(WisDevice self, WisDescriptorHeapType heap_type, WisDescriptorMemory memory_type, uint64_t size_bytes, WisDescriptorBuffer* buffer)
 {
-    return VKDeviceCreateDescriptorBuffer(self, heap_type, memory_type, descriptor_count, buffer);
+    return VKDeviceCreateDescriptorBuffer(self, heap_type, memory_type, size_bytes, buffer);
 }
 
 /**